Backend: clean up, rename package
