+++ backend/build.gradle.kts
@@ -1,17 +1,11 @@
-
-val kotlin_version: String by project
-val logback_version: String by project
-val postgres_version: String by project
-val h2_version: String by project
-
 plugins {
     kotlin("jvm") version "2.0.10"
     id("io.ktor.plugin") version "2.3.12"
     id("org.jetbrains.kotlin.plugin.serialization") version "2.0.10"
 }
 
-group = "example.com"
-version = "0.0.1"
+group = "com.github.devinkadrie.basedbin"
+version = "0.1.0"
 
 application {
     mainClass.set("io.ktor.server.netty.EngineMain")
@@ -25,16 +19,18 @@ repositories {
 }
 
 dependencies {
-    implementation("io.ktor:ktor-server-core-jvm")
-    implementation("io.ktor:ktor-server-resources-jvm")
-    implementation("io.ktor:ktor-server-cors-jvm")
-    implementation("io.ktor:ktor-serialization-kotlinx-json-jvm")
-    implementation("io.ktor:ktor-server-content-negotiation-jvm")
-    implementation("org.postgresql:postgresql:$postgres_version")
-    implementation("com.h2database:h2:$h2_version")
-    implementation("io.ktor:ktor-server-netty-jvm")
-    implementation("ch.qos.logback:logback-classic:$logback_version")
-    implementation("io.ktor:ktor-server-config-yaml")
-    testImplementation("io.ktor:ktor-server-test-host-jvm")
-    testImplementation("org.jetbrains.kotlin:kotlin-test-junit:$kotlin_version")
+    implementation("io.ktor:ktor-server-core-jvm:2.3.12")
+    implementation("io.ktor:ktor-server-resources-jvm:2.3.12")
+    implementation("io.ktor:ktor-server-cors-jvm:2.3.12")
+    implementation("io.ktor:ktor-serialization-kotlinx-json-jvm:2.3.12")
+    implementation("io.ktor:ktor-server-content-negotiation-jvm:2.3.12")
+    implementation("io.ktor:ktor-server-netty-jvm:2.3.12")
+    implementation("io.ktor:ktor-server-config-yaml:2.3.12")
+
+    implementation("org.postgresql:postgresql:42.7.2")
+    implementation("com.h2database:h2:2.1.214")
+    implementation("ch.qos.logback:logback-classic:1.4.14")
+
+    testImplementation("io.ktor:ktor-server-test-host-jvm:2.3.12")
+    testImplementation("org.jetbrains.kotlin:kotlin-test-junit:2.0.10")
 }
