+++ backend/gradle.properties
@@ -1,6 +1,1 @@
 kotlin.code.style=official
-ktor_version=2.3.12
-kotlin_version=2.0.10
-logback_version=1.4.14
-postgres_version=42.5.1
-h2_version=2.1.214
