+++ backend/src/main/kotlin/com/github/devinkadrie/basedbin/plugins/HTTP.kt
@@ -1,9 +1,8 @@
-package example.com.plugins
+package com.github.devinkadrie.basedbin.plugins
 
 import io.ktor.http.*
 import io.ktor.server.application.*
 import io.ktor.server.plugins.cors.routing.*
-import io.ktor.server.response.*
 
 fun Application.configureHTTP() {
     install(CORS) {
@@ -12,7 +11,7 @@ fun Application.configureHTTP() {
         allowMethod(HttpMethod.Delete)
         allowMethod(HttpMethod.Patch)
         allowHeader(HttpHeaders.Authorization)
-        allowHeader("MyCustomHeader")
-        anyHost() // @TODO: Don't do this in production if possible. Try to limit it.
+        // FIXME: Point at deployed host.
+        anyHost()
     }
 }
