+++ backend/src/main/kotlin/com/github/devinkadrie/basedbin/plugins/Routing.kt
@@ -1,4 +1,4 @@
-package example.com.plugins
+package com.github.devinkadrie.basedbin.plugins
 
 import io.ktor.resources.*
 import io.ktor.server.application.*
