+++ backend/src/main/kotlin/com/github/devinkadrie/basedbin/plugins/Serialization.kt
@@ -1,4 +1,4 @@
-package example.com.plugins
+package com.github.devinkadrie.basedbin.plugins
 
 import io.ktor.serialization.kotlinx.json.*
 import io.ktor.server.application.*
@@ -9,10 +9,5 @@ import io.ktor.server.routing.*
 fun Application.configureSerialization() {
     install(ContentNegotiation) {
         json()
-    }
-    routing {
-        get("/json/kotlinx-serialization") {
-                call.respond(mapOf("hello" to "world"))
-            }
     }
 }
