+++ backend/src/main/resources/application.yaml
@@ -1,6 +1,6 @@
 ktor:
     application:
         modules:
-            - example.com.ApplicationKt.module
+            - com.github.devinkadrie.basedbin.ApplicationKt.module
     deployment:
         port: 8080
